HF22 Update (Ore Mod 1.0)

+++ TFC_Shared/src/TFC/WorldGen/MapGenCaves256TFC.java
@@ -190,7 +190,7 @@ public class MapGenCaves256TFC extends MapGenBaseTFC
 											{
 												if (var50 < 10)
 												{
-													blockArray[index] = (short)Block.lavaStill.blockID;
+													blockArray[index] = (byte)Block.lavaStill.blockID;
 													metaArray[index] = 0;
 												}
 												else
